+++ lib/lang/de.dart
@@ -7,7 +7,7 @@ const Map<String, String> de = {
       'Vielen Dank!\n\nWenn Du die App unterstützen möchtest,\nfreuen wir uns auf Deine Spende ^^',
   'donationPageTitle': 'App-Entwicklung unterstützen',
   'about': 'Über die App',
-  'appVersion': 'Version 1.5',
+  'appVersion': 'Version 1.5.1',
   'record': 'Aufnehmen',
   'createMovie': 'Film erstellen',
   'settings': 'Einstellungen',
@@ -176,4 +176,12 @@ const Map<String, String> de = {
       'Migrieren aller Videos in den neuen App-Ordner, bitte warten...',
   'migrationFolderDeletionError':
       'Migration abgeschlossen, aber beim Löschen der alten Ordner ist ein Fehler aufgetreten. Bitte löschen Sie sie manuell.',
+  'preferences': 'Einstellungen',
+  'forceNativeCamera': 'Nur nativen Kamera-Modus verwenden',
+  'forceNativeCameraDescription':
+      'Verwenden Sie die native Kamera-App, um Videos aufzunehmen, anstatt die integrierte Kamera. Dies kann Probleme beim Aufzeichnen für Geräte mit geringeren Leistungsmerkmalen lösen. Diese Option funktioniert nicht auf Android 13.',
+  'movies': 'Filme',
+  'myMovies': 'Meine filme',
+  'noMoviesFound': 'Keine Filme gefunden',
+  'play': 'Abspielen',
 };