+++ lib/lang/id.dart
@@ -7,7 +7,7 @@ const Map<String, String> id = {
       'Terima kasih sudah menggunakan aplikasi ini!\n\nJika Anda ingin mendukung pembuatan aplikasi ini, jangan sungkan untuk berdonasi^^.',
   'donationPageTitle': 'Dukung pembuatan aplikasi',
   'about': 'Tentang',
-  'appVersion': 'Versi 1.5',
+  'appVersion': 'Versi 1.5.1',
   'record': 'Rekam',
   'createMovie': 'Buat video',
   'settings': 'Pengaturan',
@@ -176,4 +176,12 @@ const Map<String, String> id = {
       'Migrasi semua video Anda ke folder aplikasi baru, mohon tunggu ...',
   'migrationFolderDeletionError':
       'Migrasi selesai, tetapi terjadi kesalahan saat menghapus folder lama. Mohon hapus secara manual.',
+  'preferences': 'Preferensi',
+  'forceNativeCamera': 'Gunakan kamera bawaan',
+  'forceNativeCameraDescription':
+      'Gunakan aplikasi kamera bawaan untuk merekam video daripada kamera yang terintegrasi. Ini dapat menyelesaikan masalah merekam untuk perangkat dengan kapabilitas performa yang lebih rendah. Opsi ini tidak berfungsi pada Android 13.',
+  'movies': 'Film',
+  'myMovies': 'Film saya',
+  'noMoviesFound': 'Tidak ada film yang ditemukan',
+  'play': 'Putar',
 };