+++ lib/pages/home/daily_entry/widgets/edit_daily_button.dart
@@ -1,21 +1,44 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:image_picker/image_picker.dart';
 
 import '../../../../routes/app_pages.dart';
 import '../../../../utils/constants.dart';
 import '../../../../utils/custom_dialog.dart';
+import '../../../../utils/shared_preferences_util.dart';
 
 class EditDailyButton extends StatelessWidget {
-  void closePopupAndPushToRecording() {
+  Future<void> closePopupAndPushToRecording() async {
     Get.back();
-    Get.toNamed(Routes.RECORDING);
+
+    final sdkVersion = SharedPrefsUtil.getInt('sdkVersion');
+    final forceNativeCamera =
+        SharedPrefsUtil.getBool('forceNativeCamera') ?? false;
+    if ((sdkVersion != null && sdkVersion < 29) || forceNativeCamera) {
+      final videoFile =
+          await ImagePicker().pickVideo(source: ImageSource.camera);
+      if (videoFile != null) {
+        Get.toNamed(
+          Routes.SAVE_VIDEO,
+          arguments: {
+            'videoPath': videoFile.path,
+            'currentDate': DateTime.now(),
+            'isFromRecordingPage': true,
+          },
+        );
+      }
+    } else {
+      Get.toNamed(Routes.RECORDING);
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    return SizedBox(
-      width: MediaQuery.of(context).size.width * 0.4,
-      height: MediaQuery.of(context).size.height * 0.08,
+    return Container(
+      constraints: BoxConstraints(
+        minWidth: MediaQuery.of(context).size.width * 0.4,
+        minHeight: MediaQuery.of(context).size.height * 0.08,
+      ),
       child: ElevatedButton(
         style: ElevatedButton.styleFrom(
           backgroundColor: AppColors.purple,
@@ -34,7 +57,7 @@ class EditDailyButton extends StatelessWidget {
               content: 'editQuestion'.tr,
               actionText: 'yes'.tr,
               actionColor: AppColors.green,
-              action: () => closePopupAndPushToRecording(),
+              action: () async => await closePopupAndPushToRecording(),
               action2Text: 'no'.tr,
               action2Color: Colors.red,
               action2: () => Get.back(),
@@ -42,13 +65,14 @@ class EditDailyButton extends StatelessWidget {
           );
         },
         child: Row(
-          mainAxisAlignment: MainAxisAlignment.spaceEvenly,
+          mainAxisSize: MainAxisSize.min,
           children: [
             Icon(
               Icons.edit,
               size: MediaQuery.of(context).size.width * 0.07,
               color: Colors.white,
             ),
+            const SizedBox(width: 15.0),
             Text(
               'edit'.tr,
               style: TextStyle(