+++ CHANGELOG.md
@@ -1,8 +1,17 @@
+## v1.5.1 - 02/2023
+- Improved movies viewer page ordering movies by most recent
+- Fixed subtitles text field not recognizing some cyrillic characters
+- Fixed edit button not wrapping text correctly in some languages
+- Fixed saving videos for profiles with white spaces in the name
+- Fixed error when saving videos due missing audio permission in some devices
+- Fixed UI in save recording page not updating after a custom locale is set twice
+
 ## v1.5 - 01/2023
 - New features:
     - Upload from gallery
     - Vertical video support
     - Calendar page
+    - Movies viewer page
     - Profiles
     - Subtitles
     - Video trimming
@@ -20,6 +29,7 @@
     - Routine to save logs in txt files
     - Recording page now supports portrait mode, zoom and focus controls
     - Date editing preferences are now saved
+    - Added option to use native camera for in-app video recording
 
 ## v1.1 - 05/2021
 - Added feature to receive daily notifications